+++ cookbooks/aws-parallelcluster-shared/metadata.rb
@@ -16,4 +16,4 @@ supports 'redhat', '= 8.7'
 
 depends 'pyenv', '~> 4.2.3'
 depends 'yum', '~> 7.4.13'
-depends 'yum-epel', '~> 4.5.0'
+depends 'yum-epel', '~> 5.0.2'