+++ cookbooks/aws-parallelcluster-slurm/metadata.rb
@@ -13,7 +13,7 @@ supports 'amazon', '>= 2.0'
 supports 'centos', '>= 7.0'
 supports 'ubuntu', '>= 20.04'
 
-depends 'apt', '~> 7.4.2'
+depends 'apt', '~> 7.5.14'
 depends 'iptables', '~> 8.0.0'
 depends 'line', '~> 4.5.2'
 depends 'nfs', '~> 5.0.0'