+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
   - openssh-2.11.3 (from openssh-2.10.3)
   - pyenv-4.2.3 (from pyenv-3.5.1)
   - selinux-6.1.12 (from selinux-6.0.5)
+  - yum-7.4.13 (from yum-7.4.0)
 
 **BUG FIXES**