+++ CHANGELOG.md
@@ -3,6 +3,18 @@ aws-parallelcluster-cookbook CHANGELOG
 
 This file is used to list changes made in each version of the AWS ParallelCluster cookbook.
 
+3.8.0
+------
+
+**ENHANCEMENTS**
+
+**CHANGES**
+- Upgrade third-party cookbook dependencies:
+  - apt-7.5.14 (from apt-7.4.0)
+  - line-4.5.13 (from line-4.5.2)
+
+**BUG FIXES**
+
 3.7.0
 ------
 
@@ -33,6 +45,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Upgrade third-party cookbook dependencies:
     - apt-7.5.14 (from apt-7.4.0)
     - line-4.5.13 (from line-4.5.2)
+    - openssh-2.11.3 (from openssh-2.10.3)
 
 **BUG FIXES**
 - Fix cluster creation failure with Ubuntu Deep Learning AMI on GPU instances and DCV enabled.